+++ docs/README.txt
@@ -1,0 +1,66 @@
+(eng)------------------------------------------------------------------------------------------------------
+
+Project "Forming unique words from text"
+ + translation of words from English to Russian
+
+Documentation in ./docs folder
+
+python 3.11.5
+
+There are two files in the project:
+- input.txt is a file with text, on the basis of which the output file of unique words is formed.
+- output.txt is an output file with unique words. Each word is in a separate line.
+
+When running the py-file, the user is asked to:
+- To fomize the output file with unique words with or without translation.
+
+
+To deploy the project:
+	1. Create a working environment
+	   
+	   python3 -m venv venv          (Linux/MacOS)
+	   python -m venv venv           (Windows)
+	   
+	2 Activate ENV
+
+	   source venv/bin/activate      (Linux/macOS)
+	   .\venv\Scripts\Activate       (Windows) 
+
+	3. Download the required libraries
+
+	   pip install -r requirements.txt   
+
+
+(rus)------------------------------------------------------------------------------------------------------
+
+Проект "Формирование уникальных слов из текста"
+ + перевод слов с английского на русский
+
+Документация в папке ./docs
+
+python 3.11.5
+
+В проекте два файла:
+- input.txt  это файл с текстом, на основании которого формируется выходной файл уникальных слов.
+- output.txt выходной файл с уникальными словами. Каждое слово в отдельной строке.
+
+При запуске файла, у пользователя запрашивается:
+- Фомировать выходной файл с уникальными словами с переводом или без.
+
+
+Равернуть проект:
+1. Создать рабочую среду
+   
+   python3 -m venv venv (Linux/MacOS)
+   python -m venv venv (Windows)
+   
+2. Активировать ENV
+
+   source venv/bin/activate  (Linux/macOS)
+
+   .\venv\Scripts\Activate (Windows) 
+
+3. Подгрузить необходимые библиотеки
+
+   pip install -r requirements.txt   
+
